Don't use GetFileSize any more
It does not return INVALID_FILE_SIZE on large files. Hence, limit file sizes to <INT_MAX.

+++ src/TortoiseProc/gitlogcache.h
@@ -113,7 +113,7 @@ protected:
 	HANDLE m_DataFile = INVALID_HANDLE_VALUE;
 	HANDLE m_DataFileMap = nullptr;
 	BYTE* m_pCacheData = nullptr;
-	DWORD m_DataFileLength = 0;
+	size_t m_DataFileLength = 0;
 
 	void CloseDataHandles();
 	void CloseIndexHandles();
